+++ source/codegen/metadata/nifake/functions.py
@@ -2521,5 +2521,71 @@ functions = {
             }
         ],
         'returns': 'ViStatus'
+    },
+    'GetViUInt8': {
+        'codegen_method': 'public',
+        'parameters': [
+            {
+                'direction': 'in',
+                'name': 'vi',
+                'type': 'ViSession'
+            },
+            {
+                'direction': 'out',
+                'name': 'aUint8Number',
+                'type': 'ViUInt8'
+            }
+        ],
+        'returns': 'ViStatus'
+    },
+    'ViUInt8ArrayInputFunction': {
+        'codegen_method': 'public',
+        'parameters': [
+            {
+                'direction': 'in',
+                'name': 'vi',
+                'type': 'ViSession'
+            },
+            {
+                'direction': 'in',
+                'name': 'numberOfElements',
+                'type': 'ViInt32'
+            },
+            {
+                'direction': 'in',
+                'name': 'anArray',
+                'size': {
+                    'mechanism': 'passed-in',
+                    'value': 'numberOfElements'
+                },
+                'type': 'ViUInt8[]'
+            }
+        ],
+        'returns': 'ViStatus'
+    },
+    'ViUInt8ArrayOutputFunction': {
+        'codegen_method': 'public',
+        'parameters': [
+            {
+                'direction': 'in',
+                'name': 'vi',
+                'type': 'ViSession'
+            },
+            {
+                'direction': 'in',
+                'name': 'numberOfElements',
+                'type': 'ViInt32'
+            },
+            {
+                'direction': 'out',
+                'name': 'anArray',
+                'size': {
+                    'mechanism': 'passed-in',
+                    'value': 'numberOfElements'
+                },
+                'type': 'ViUInt8[]'
+            }
+        ],
+        'returns': 'ViStatus'
     }
 }